data/init_postgres_requirements.txt
psycopg2-binary
python-dotenv
pandas